Add files via upload
In this assignment, I applied the event listener functions to "index.html".

Firstly, the images in the middle of the home page are of essential commercially value. In order to utilize this space, I add event listener functions to the images, which redirect users to the pages of the product and game.

Secondly, according to the feedback of the usability test, one of the most important things on the website is the navigation bar. Therefore, in order to make the buttons displayed in the navigation bar more appealing to users, the event listener functions are placed to track users' actions.

And in the A/B test, I plan to change the sequence of the "周边" button in the navigation bar, which is of commercial value. For instance, if the "周边" button is moved to the front of the "测评" button, will click ratio of the "周边" button increase?